Restore focus after _detachColumns() + _reattachColumns()
Previously when _detachColumns() removed the DOM element that currently had the focus (i.e. the document.activeElement), then the focus fell to the <body> of the current document and was left there, preventing keyboard events from being fired. To fix this, I now save the current activeElement and restore it in _reattachColumns().